ゲームプレイとタイトルシーンの更新
* `Enemy.cpp` の `Enemy::Update` メソッドで、`M_PI` を `static_cast<float>(M_PI)` にキャストする変更が行われました。
* `Player.cpp` の複数のメソッドで、攻撃や回避、衝突処理の改善が行われました。
* `PlayerWepon.cpp` の衝突処理メソッドで、引数 `other` を使用しないことを明示するためのコードが追加されました。
* `Line.cpp` の複数のメソッドで、引数を `static_cast<UINT>` でキャストする変更が行われました。
* `LineManager.cpp` の複数のメソッドで、デバッグ用のコードがコメントアウトされました。
* `CameraManager.cpp` の複数のメソッドで、初期カメラ設定やデストラクタの削除が行われました。
* `SceneManager.cpp` の `SceneManager::Initialize` メソッドで、初期シーンを `TitleScene` に変更する変更が行われました。
* `GamePlayScene.cpp` の複数のメソッドで、ゲームクリア処理やスプライトの描画処理が追加されました。
* `GamePlayScene.h` にゲームクリアスプライトと関連するメンバー変数が追加されました。
* `TitleScene.cpp` の複数のメソッドで、音の読み込みやタイトルオブジェクト、スプライトの初期化と更新処理が追加されました。
* `TitleScene.h` に新しいメンバ変数が追加されました。

+++ application/Enemy.cpp
@@ -50,9 +50,9 @@ void Enemy::Update(const Vector3 &playerPos) {
 			//放射状に線を描画してエフェクトを出す
 		float currentLength = maxLength * ( frameCount / 60.0f ); // 60フレームで最大長に達する
 		for(int i = 0; i < numLines; ++i) {
-			float angle = ( 2.0f * M_PI / numLines ) * i;
+			float angle = ( 2.0f * static_cast<float>( M_PI ) / numLines ) * i;
 			for(int j = -1; j <= 1; ++j) {
-				float verticalAngle = ( M_PI / 2.0f ) * j;
+				float verticalAngle = ( static_cast<float>( M_PI ) / 2.0f ) * j;
 				Vector3 direction = { cos(angle) * cos(verticalAngle), sin(verticalAngle), sin(angle) * cos(verticalAngle) };
 				Vector3 endPosition = transform.translate + direction * currentLength;
 				LineManager::GetInstance()->DrawLine(transform.translate, endPosition, { 1.0f, 0.0f, 0.0f });
